CH15/ch15-1/langserveapp/requirements.txt
qdrant-client>=1.10.1,<2.0.0
python-dotenv>=1.0.1,<2.0.0
langchain-core==0.3.0
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.1
langchain-qdrant==0.1.4
langsmith>=0.1.112,<0.2.0
numpy>=1.26.0,<2.0.0
langserve[server]==0.3.0
httpx>=0.23.0,<0.28.0
pydantic>2
uvicorn>=0.23.2
pypdf
